# Improved loop unrolling