--- a/popcnt-avx2-lookup.cpp
+++ b/popcnt-avx2-lookup.cpp
@@ -18,25 +18,33 @@ std::uint64_t popcnt_AVX2_lookup(const uint8_t* data, const size_t n) {
 
     __m256i acc = _mm256_setzero_si256();
 
-    while (i + 32 < n) {
+#define ITER { \
+        const __m256i vec = _mm256_loadu_si256(reinterpret_cast<const __m256i*>(data + i)); \
+        const __m256i lo  = _mm256_and_si256(vec, low_mask); \
+        const __m256i hi  = _mm256_and_si256(_mm256_srli_epi16(vec, 4), low_mask); \
+        const __m256i popcnt1 = _mm256_shuffle_epi8(lookup, lo); \
+        const __m256i popcnt2 = _mm256_shuffle_epi8(lookup, hi); \
+        local = _mm256_add_epi8(local, popcnt1); \
+        local = _mm256_add_epi8(local, popcnt2); \
+        i += 32; \
+    }
 
-        __m256i local = _mm256_setzero_si256(); 
-
-        for (int k=0; k < 255/8 && i + 32 < n; k++, i += 32) {
-            const __m256i vec = _mm256_loadu_si256(reinterpret_cast<const __m256i*>(data + i));
-            const __m256i lo  = _mm256_and_si256(vec, low_mask);
-            const __m256i hi  = _mm256_and_si256(_mm256_srli_epi16(vec, 4), low_mask);
-
-            const __m256i popcnt1 = _mm256_shuffle_epi8(lookup, lo);
-            const __m256i popcnt2 = _mm256_shuffle_epi8(lookup, hi);
-
-            local = _mm256_add_epi8(local, popcnt1);
-            local = _mm256_add_epi8(local, popcnt2);
-        }
-
+    while (i + 8*32 <= n) {
+        __m256i local = _mm256_setzero_si256();
+        ITER ITER ITER ITER
+        ITER ITER ITER ITER
         acc = _mm256_add_epi64(acc, _mm256_sad_epu8(local, _mm256_setzero_si256()));
     }
 
+    __m256i local = _mm256_setzero_si256();
+
+    while (i + 32 <= n) {
+        ITER;
+    }
+
+    acc = _mm256_add_epi64(acc, _mm256_sad_epu8(local, _mm256_setzero_si256()));
+
+#undef ITER
 
     uint64_t result = 0;