--- a/popcnt-cpu.cpp
+++ b/popcnt-cpu.cpp
@@ -2,14 +2,22 @@ std::uint64_t popcnt_cpu_64bit(const uint8_t* data, const size_t n) {
 
     uint64_t result = 0;
 
-    for (size_t i=0; i < n; i += 8) {
-        const uint64_t v = *reinterpret_cast<const uint64_t*>(data + i);
-
-        result += _popcnt64(v);
+    uint64_t v, i = 0;
+#define ITER { \
+        v = *reinterpret_cast<const uint64_t*>(data + i); \
+        result += _popcnt64(v); \
+        i += 8; \
     }
 
-    for (size_t i=8*(n/8); i < n; i++) {
+    while (i + 4*8 <= n) {
+        ITER ITER ITER ITER
+    }
+
+#undef ITER
+
+    while (i < n) {
         result += lookup8bit[data[i]];
+        i++;
     }
 
     return result;