Adds loop unrolling mostly everywhere, improving each version's running time by ~30%.

Before:
```
$ ./speed_avx2 10000 1000000
LUT (uint8_t[256])            ... time = 5.239423 s
LUT (uint64_t[256])           ... time = 5.792041 s (speedup: 0.90)
bit parallel                  ... time = 5.328451 s (speedup: 0.98)
bit parallel optimized        ... time = 3.660314 s (speedup: 1.43)
bit parallel optimized - SSE  ... time = 1.425489 s (speedup: 3.68)
SSSE3 [pshufb]                ... time = 0.977139 s (speedup: 5.36)
AVX2 [pshufb]                 ... time = 0.534345 s (speedup: 9.81)
CPU popcnt                    ... time = 0.773064 s (speedup: 6.78)
```

After:
```
$ ./speed_avx2 10000 1000000
LUT (uint8_t[256])            ... time = 4.322537 s
LUT (uint64_t[256])           ... time = 4.279465 s (speedup: 1.01)
bit parallel                  ... time = 4.223320 s (speedup: 1.02)
bit parallel optimized        ... time = 2.698201 s (speedup: 1.60)
bit parallel optimized - SSE  ... time = 1.241662 s (speedup: 3.48)
SSSE3 [pshufb]                ... time = 0.666053 s (speedup: 6.49)
AVX2 [pshufb]                 ... time = 0.352794 s (speedup: 12.25)
CPU popcnt                    ... time = 0.518836 s (speedup: 8.33)
```