@WojciechMula 2016-03-13
Thanks a lot! I completely forget about manual unrolling.